+++ src/admin/handlers/collections/item/edit_form.rs
@@ -156,6 +156,7 @@ fn prepare_edit_fields(
             .filter_hidden(true)
             .non_default_locale(non_default_locale)
             .doc_id(Some(id))
+            .user(get_user_doc(auth_user))
             .build(),
     );